Address comments from jinxing64
